:bug: :: 크라우드 펀딩 남은 날짜 계산 fix

+++ crowdfunding/src/main/kotlin/com/project/indistraw/crowdfunding/application/port/input/dto/CrowdfundingDetailDto.kt
@@ -8,7 +8,7 @@ data class CrowdfundingDetailDto(
     val description: String,
     val writer: Writer,
     val amount: Amount,
-    val remainingDay: Int,
+    val remainingDay: Long,
     val fundingCount: Long,
     val reward: List<RewardDto>,
     val statusType: StatusType,
